Add files via upload
WEB SCRAPING – ASSIGNMENT 1
In all the following questions, you have to use BeautifulSoup to scrape different websites and collect data as per
the requirement of the question.
Every answer to the question should be in form of a python function which should take URL as the parameter.
Use Jupyter Notebooks to program, upload it on your GitHub and send the link of the jupyter notebook to your
SME.
1. Write a python program to display all the header tags from ‘en.wikipedia.org/wiki/Main_Page’.
2. Write a python program to display IMDB’s Top rated 100 movies’ data (i.e. Name, IMDB rating, Year of
release) and save it in form of a CSV file.
3. Write a python program to display IMDB’s Top rated 100 Indian movies’ data (i.e. Name, IMDB rating, Year
of release) and save it in form of a CSV file.
4. Write a python program to scrap book name, author name, genre and book review of any 5 books from
‘www.bookpage.com’
5. Write a python program to scrape cricket rankings from ‘www.icc-cricket.com’. You have to scrape:
i) Top 10 ODI teams in men’s cricket along with the records for matches, points and rating.
ii) Top 10 ODI Batsmen in men along with the records of their team and rating.
iii) Top 10 ODI bowlers along with the records of their team and rating.
6. Write a python program to scrape cricket rankings from ‘www.icc-cricket.com’. You have to scrape:
i) Top 10 ODI teams in women’s cricket along with the records for matches, points, and rating.
ii) Top 10 women’s ODI players along with the records of their team and rating.
iii) Top 10 women’s ODI all-rounder along with the records of their team and rating.
7. Write a python program to scrape details of all the mobile phones under Rs. 20,000 listed on Amazon.in. The
scraped data should include Product Name, Price, Image URL, and Average Rating.
8. Write a python program to extract information about the local weather from the National Weather Service
website of USA, https://www.weather.gov/ for the city, San Francisco. You need to extract data about 7 days
extended forecast display for the city. The data should include a period, short description, temperature, and
description.