News Classification
News archives contain several Gigabytes of data, which is not easy to handle. For simplicity they are stored in simplest format possible. But during analysis, analysts require a meta information about each news article. Sometimes it is available but there are cases where meta information is missing. This is where data science comes into play. Use cases are built to analyse the available meta data and hence analyse meta information about the news articles, where it is not present. In this project we have to build such use case where we can generate the metadata around news articles and make it ready for further analysis. In phase 1, we need to gather information from archives of past few years from different news websites. In this phase we will scrape the news websites for news and related information.
The task assigned to you is to scrape archives from The Times of India website for Feb, 2021. You will be allotted tasks to scrape data based on different timeframes. The details to be scraped are given below:
1. Date: This column will contain the date on which news was reported.
2. Author: This column will contain name of the author of news report.
3. Vertical: This column will contain the news verticals (e.g. Entertainment, Sports, Lifestyle, Economy, etc.).
4. Headline: This column will contain the news headline.
5. Description: This column will contain the main content of the news report.
You have to write a code to scrape the above data from The Times of India website. The time frames of the archives will be allocated to you individually.