Ratings Prediction Project
We have a client who has a website where people write different reviews for technical 
products. Now they are adding a new feature to their website i.e. The reviewer will have to 
add stars(rating) as well with the review. The rating is out 5 stars and it only has 5 options 
available 1 star, 2 stars, 3 stars, 4 stars, 5 stars. Now they want to predict ratings for the 
reviews which were written in the past and they don’t have a rating. So, we have to build an 
application which can predict the rating by seeing the review.